Made changes to the finders so that a search like :conditions => { :id => 10 } will work. Added TimeConverter to go with Date and DateTime and moved the config defaults into the Config module Got ActiveModel::Dirty working again as best I could.  Not sure what happened there.
Making sure that #attributes returns the Ruby-converted versions of each property

Added attribute_will_change! back in there for ActiveModel::Dirty

Fixed the method_missing so it causes an exception when it's really missing a method.  Added a bit to get properties working when they haven't got methods defined, mainly for ActiveModel::Dirty.

Changed avator to avatar

Added a test for changing properties after a model has been saved.  Right now, assignment to an attribute changes the attribute in the DB instead of waiting for a save so that validations and callbacks can be run.  Will need some major refactoring to get that to work well though, if we want to be able to assign relationships to unsaved models as well...

First stab at handling a special case for find, where :id is used as a condition

Added TypeConverter defaults in here

Added #converters class method

Added Time converter