Implemented paging using the will_paginate gem.
Now includes the will_paginate gem in the gemspec.
Paginating is now exposed in the Neo4j::NodeMixin#pageinate (which is included in the Neo4j::Rails::Model)
Have not tested this with rails, but it might already work.
TODO 1: currently only lucene queries are supported to paginate, should be possible to paginate traversals as well.
TODO 2: the full query api is not exposed using the paginate method, need at least be able to sort.

Maybe should change the query API and intead rely on hashes. We now use a DSL for specifying ranges and sort orders which
 might not be possible to combine with paginate which (I think) expects the full query in an hash.

 Example:
   paginate(:sort_by => ''asc:age", :query => 'query',) vs.  find(query).asc(:age)