fix the issue with serializing String values
Before, any non-String value was appropriately to_json'd in the to_db and retrieved back out
via the to_ruby.

String would break because the underlying JSON db_type was String _already_, so it did not need
to be converted (original code) but would blow up when it tried.

This fixes that by assuming that all values need to be to_json going in and from_json going out.