Update Dockerfile with current "activegraph" naming
The gem name was changed from `neo4j` to `activegraph` and thus the
folder structure changed as well. Update the Dockerfile to keep in
line with this change.

Follow Github Action's lead and install bundler version pessimistically
to latest version of 2.x

Activegraph rspec suites use the bolt adapter these days which has a
dependency on the Seabolt package. Install the latest package version
for Ubuntu/debian in the Dockerfile.